requirements.txt
# ── Core ──────────────────────────────────────
requests==2.31.0

# ── Testing Framework ─────────────────────────
pytest==8.1.0
pytest-html==4.1.1
pytest-asyncio==0.21.0
pytest-repeat==0.9.3
pytest-xdist==3.5.0

# ── LLM Observability ─────────────────────────
langsmith==0.7.5
python-dotenv==1.2.1
langchain==0.1.20

# ── LLM Evaluation ────────────────────────────
deepeval==1.4.3

#- Consistsny check for LLM responses
pytest-repeat==0.9.4 

sentence-transformers==5.2.3
pandas==3.0.1
numpy==1.26.4
